backend/app-agricola-backend/app/sales/api/views.py
from rest_framework import generics
from sales.models import Sale
from sales.api.serializers import SaleSerializer

class SaleListCreateAPIView(generics.ListCreateAPIView):
    queryset = Sale.objects.all()
    serializer_class = SaleSerializer

class SaleDetailAPIView(generics.RetrieveUpdateDestroyAPIView):
    queryset = Sale.objects.all()
    serializer_class = SaleSerializer
